+++ app/Http/Controllers/Api/CategorieController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\User;
 use App\Repositories\CategorieRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class CategorieController extends Controller
@@ -24,18 +26,27 @@ class CategorieController extends Controller
     public function index()
     {
         //
-        $categories=$this->categorieRepository->findAll();
-        if($categories){
+        if(Auth::user()->can('viewAny',User::class)){
+            $categories=$this->categorieRepository->findAll();
+            if($categories){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$categories,
+                    "message"=>"Liste des Catégories"
+                ],Response::HTTP_FOUND);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$categories,
-                "message"=>"Liste des Catégories"
-            ],Response::HTTP_FOUND);
-        }
-        return response()->json([
                 "sucess"=>false,
                 "message"=>"Pas de catégorie trouvé"
-        ],Response::HTTP_NOT_FOUND);
+            ],Response::HTTP_NOT_FOUND);
+        }
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas de autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -53,18 +64,27 @@ class CategorieController extends Controller
     public function store(Request $request)
     {
         //
-        $categorie=$this->categorieRepository->create($request->all());
-        if($categorie){
+        if(Auth::user()->can('create',User::class)){
+            $categorie=$this->categorieRepository->create($request->all());
+            if($categorie){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$categorie,
+                    "message"=>"Catégorie inserée"
+                ],Response::HTTP_CREATED);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$categorie,
-                "message"=>"Catégorie inserée"
-            ],Response::HTTP_CREATED);
+                "sucess"=>false,
+                "message"=>"Erreur lors de l'insertion d'une catégorie"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur lors de l'insertion d'une catégorie"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -73,18 +93,27 @@ class CategorieController extends Controller
     public function show($id)
     {
         //
-        $categorie=$this->categorieRepository->findById($id);
-        if($categorie){
+        if(Auth::user()->can('view',User::class)){
+            $categorie=$this->categorieRepository->findById($id);
+            if($categorie){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$categorie,
+                    "message"=>"Catégorie trouvée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$categorie,
-                "message"=>"Catégorie trouvée"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Catégorie inexistante"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Catégorie inexistante"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -101,18 +130,27 @@ class CategorieController extends Controller
     public function update(Request $request, $id)
     {
         //
-        $categorie=$this->categorieRepository->update($request->all(),$id);
-        if($categorie){
+        if(Auth::user()->can('update',User::class)){
+            $categorie=$this->categorieRepository->update($request->all(),$id);
+            if($categorie){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$categorie,
+                    "message"=>"Mise à jour effectuée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$categorie,
-                "message"=>"Mise à jour effectuée"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Erreur survenue lors de la mise à jour"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur survenue lors de la mise à jour"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -121,16 +159,25 @@ class CategorieController extends Controller
     public function destroy($id)
     {
         //
-        $categorie=$this->categorieRepository->delete($id);
-        if($categorie>0){
+        if(Auth::user()->can('delete',User::class)){
+            $categorie=$this->categorieRepository->delete($id);
+            if($categorie>0){
+                return response()->json([
+                    "sucess"=>true,
+                    "message"=>"Suppression réussie"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "message"=>"Suppression réussie"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Une erreur s'est produite..."
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Une erreur s'est produite..."
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 }
