+++ app/Http/Controllers/Api/RoleController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Repositories\RoleRepository;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
@@ -25,18 +27,27 @@ class RoleController extends Controller
     public function index()
     {
         //
-        $roles=$this->roleRepository->findAll();
-        if($roles){
+        if(Auth::user()->can('viewAny',User::class)){
+            $roles=$this->roleRepository->findAll();
+            if($roles){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$roles,
+                    "message"=>"Liste des rôles"
+                ],Response::HTTP_FOUND);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$roles,
-                "message"=>"Liste des rôles"
-            ],Response::HTTP_FOUND);
+                "sucess"=>false,
+                "message"=>"Pas de rôle trouvé"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Pas de rôle trouvé"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
 
     }
 
@@ -54,18 +65,27 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         //
-        $role=$this->roleRepository->create($request->all());
-        if($role){
+        if(Auth::user()->can('create',User::class)){
+            $role=$this->roleRepository->create($request->all());
+            if($role){
+                return response()->json([
+                    "success"=>true,
+                    "data"=>$role,
+                    "message"=>"Rôle inseré"
+                ],Response::HTTP_CREATED);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$role,
-                "message"=>"Rôle inseré"
-            ],Response::HTTP_CREATED);
+                "success"=>false,
+                "message"=>"Erreur lors de l'insertion d'un rôle"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur lors de l'insertion d'un rôle"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "success"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_NOT_FOUND);
+        }
+
     }
 
     /**
@@ -74,18 +94,27 @@ class RoleController extends Controller
     public function show($id)
     {
         //
-        $role=$this->roleRepository->findById($id);
-        if($role){
+        if(Auth::user()->can('view',User::class)){
+            $role=$this->roleRepository->findById($id);
+            if($role){
+                return response()->json([
+                    "success"=>true,
+                    "data"=>$role,
+                    "message"=>"Rôle trouvé"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$role,
-                "message"=>"Rôle trouvé"
-            ],Response::HTTP_OK);
+                "success"=>false,
+                "message"=>"Rôle inexistant"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Rôle inexistant"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "success"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
 
     }
 
@@ -103,18 +132,27 @@ class RoleController extends Controller
     public function update(Request $request,  $id)
     {
         //
-        $role=$this->roleRepository->update($request->except(['id']),$id);
-        if($role){
+        if(Auth::user()->can('update',User::class)){
+            $role=$this->roleRepository->update($request->except(['id']),$id);
+            if($role){
+                return response()->json([
+                    "success"=>true,
+                    "data"=>$role,
+                    "message"=>"Mise à jour effectuée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$role,
-                "message"=>"Mise à jour effectuée"
-            ],Response::HTTP_OK);
+                "success"=>false,
+                "message"=>"Erreur survenue lors de la mise à jour"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur survenue lors de la mise à jour"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "success"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -123,16 +161,25 @@ class RoleController extends Controller
     public function destroy($id)
     {
         //
-        $role=$this->roleRepository->delete($id);
-        if($role>0){
+        if(Auth::user()->can('delete',User::class)){
+            $role=$this->roleRepository->delete($id);
+            if($role>0){
+                return response()->json([
+                    "success"=>true,
+                    "message"=>"Suppression réussie"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "message"=>"Suppression réussie"
-            ],Response::HTTP_OK);
+                "success"=>false,
+                "message"=>"Une erreur s'est produite..."
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Une erreur s'est produite..."
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "success"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 }
