+++ app/Http/Controllers/Api/RubriqueController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\User;
 use App\Repositories\RubriqueRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class RubriqueController extends Controller
@@ -24,18 +26,27 @@ class RubriqueController extends Controller
     public function index()
     {
         //
-        $rubriques=$this->rubriquerepository->findAll();
-        if($rubriques){
+        if(Auth::user()->can('viewAny',User::class)){
+            $rubriques=$this->rubriquerepository->findAll();
+            if($rubriques){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$rubriques,
+                    "message"=>"Liste des rubriques"
+                ],Response::HTTP_FOUND);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$rubriques,
-                "message"=>"Liste des rubriques"
-            ],Response::HTTP_FOUND);
+                "sucess"=>false,
+                "message"=>"Pas de rubrique trouvée"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Pas de rubrique trouvée"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -52,19 +63,27 @@ class RubriqueController extends Controller
     public function store(Request $request)
     {
         //
-
-        $rubrique=$this->rubriquerepository->create($request->all());
-        if($rubrique){
+        if(Auth::user()->can('create',User::class)){
+            $rubrique=$this->rubriquerepository->create($request->all());
+            if($rubrique){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$rubrique,
+                    "message"=>"Rubrique inserée"
+                ],Response::HTTP_CREATED);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$rubrique,
-                "message"=>"Rubrique inserée"
-            ],Response::HTTP_CREATED);
+                "sucess"=>false,
+                "message"=>"Erreur lors de l'insertion d'une rubrique"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur lors de l'insertion d'une rubrique"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -73,18 +92,27 @@ class RubriqueController extends Controller
     public function show($id)
     {
         //
-        $rubrique=$this->rubriquerepository->findById($id);
-        if($rubrique){
+        if(Auth::user()->can('view',User::class)){
+            $rubrique=$this->rubriquerepository->findById($id);
+            if($rubrique){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$rubrique,
+                    "message"=>"Rubrique trouvée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$rubrique,
-                "message"=>"Rubrique trouvée"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Rubrique inexistante"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Rubrique inexistante"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -101,18 +129,27 @@ class RubriqueController extends Controller
     public function update(Request $request,  $id)
     {
         //
-        $rubrique=$this->rubriquerepository->update($request->except(['id']),$id);
-        if($rubrique){
+        if(Auth::user()->can('update',User::class)){
+            $rubrique=$this->rubriquerepository->update($request->except(['id']),$id);
+            if($rubrique){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$rubrique,
+                    "message"=>"Mise à jour effectuée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$rubrique,
-                "message"=>"Mise à jour effectuée"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Erreur survenue lors de la mise à jour"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur survenue lors de la mise à jour"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -121,16 +158,25 @@ class RubriqueController extends Controller
     public function destroy($id)
     {
         //
-        $rubrique=$this->rubriquerepository->delete($id);
-        if($rubrique>0){
+        if(Auth::user()->can('delete',User::class,Role::class)){
+            $rubrique=$this->rubriquerepository->delete($id);
+            if($rubrique>0){
+                return response()->json([
+                    "sucess"=>true,
+                    "message"=>"Suppression réussie"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "message"=>"Suppression réussie"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Une erreur s'est produite..."
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Une erreur s'est produite..."
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 }
