+++ app/Http/Controllers/Api/TypePubController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Typepub;
+use App\Models\User;
 use App\Repositories\TypePubRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class TypePubController extends Controller
@@ -25,18 +27,27 @@ class TypePubController extends Controller
     public function index()
     {
         //
-        $typepub=$this->typepubRepository->findAll();
-        if($typepub){
+        if(Auth::user()->can('viewAny',User::class)){
+            $typepub=$this->typepubRepository->findAll();
+            if($typepub){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$typepub,
+                    "message"=>"Liste des TypePubs"
+                ],Response::HTTP_FOUND);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$typepub,
-                "message"=>"Liste des TypePubs"
-            ],Response::HTTP_FOUND);
+                "sucess"=>false,
+                "message"=>"Pas de typepub trouvé"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Pas de typepub trouvé"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -53,18 +64,27 @@ class TypePubController extends Controller
     public function store(Request $request)
     {
         //
-        $typepub=$this->typepubRepository->create($request->all());
-        if($typepub){
+        if(Auth::user()->can('create',User::class)){
+            $typepub=$this->typepubRepository->create($request->all());
+            if($typepub){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$typepub,
+                    "message"=>"TypPub inseré"
+                ],Response::HTTP_CREATED);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$typepub,
-                "message"=>"TypPub inseré"
-            ],Response::HTTP_CREATED);
+                "sucess"=>false,
+                "message"=>"Erreur lors de l'insertion d'un typepub"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur lors de l'insertion d'un typepub"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -73,18 +93,27 @@ class TypePubController extends Controller
     public function show( $id)
     {
         //
-        $typepub=$this->typepubRepository->findById($id);
-        if($typepub){
+        if(Auth::user()->can('view',User::class)){
+            $typepub=$this->typepubRepository->findById($id);
+            if($typepub){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$typepub,
+                    "message"=>"TypePub trouvé"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$typepub,
-                "message"=>"TypePub trouvé"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"TypePub inexistant"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"TypePub inexistant"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -101,18 +130,27 @@ class TypePubController extends Controller
     public function update(Request $request, $id)
     {
         //
-        $typepub=$this->typepubRepository->update($request->all(),$id);
-        if($typepub){
+        if(Auth::user()->can('update',User::class)){
+            $typepub=$this->typepubRepository->update($request->all(),$id);
+            if($typepub){
+                return response()->json([
+                    "sucess"=>true,
+                    "data"=>$typepub,
+                    "message"=>"Mise à jour effectuée"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "data"=>$typepub,
-                "message"=>"Mise à jour effectuée"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Erreur survenue lors de la mise à jour"
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Erreur survenue lors de la mise à jour"
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -121,16 +159,25 @@ class TypePubController extends Controller
     public function destroy($id)
     {
         //
-        $typepub=$this->typepubRepository->delete($id);
-        if($typepub>0){
+        if(Auth::user()->can('delete',User::class)){
+            $typepub=$this->typepubRepository->delete($id);
+            if($typepub>0){
+                return response()->json([
+                    "sucess"=>true,
+                    "message"=>"Suppression réussie"
+                ],Response::HTTP_OK);
+            }
             return response()->json([
-                "sucess"=>true,
-                "message"=>"Suppression réussie"
-            ],Response::HTTP_OK);
+                "sucess"=>false,
+                "message"=>"Une erreur s'est produite..."
+            ],Response::HTTP_NOT_FOUND);
         }
-        return response()->json([
-            "sucess"=>false,
-            "message"=>"Une erreur s'est produite..."
-        ],Response::HTTP_NOT_FOUND);
+        else{
+            return response()->json([
+                "sucess"=>false,
+                "message"=>"Pas autorisé"
+            ],Response::HTTP_UNAUTHORIZED);
+        }
+
     }
 }
